+++ packages/react-charts/src/types/font.ts
@@ -1,0 +1,7 @@
+export type FontWeight = 'regular' | 'bold' | 'black';
+
+export type FontFamily = {
+  regular: string;
+  bold: string;
+  black: string;
+};