+++ packages/react-core/src/components/atoms/Text/Text.tsx
@@ -30,6 +30,7 @@ export interface TextProps {
   capitalFirst?: boolean;
   style?: StyleProp<TextStyle>;
   children?: React.ReactNode;
+  testID?: string;
 }
 
 /** NOTE: When using colors, be careful to not override fontColor by using colorVariant and colorTone, referent to theme colors and not text colors. */